backend/accounts: add type for OrderedTransactions
Refactor btc and eth to not repeat the same work of sorting.

In a later commit, the balance at each tx will be computed there as
well, so it can be used in the portfolio overview.
