--- backend/coins/eth/sort.go
@@ -1,27 +1,0 @@
-// Copyright 2020 Shift Devices AG
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package eth
-
-import "github.com/digitalbitbox/bitbox-wallet-app/backend/accounts"
-
-// byHeight defines the methods needed to satisify sort.Interface to sort transactions by their
-// confirmations.
-type byHeight []*accounts.TransactionData
-
-func (s byHeight) Len() int { return len(s) }
-func (s byHeight) Less(i, j int) bool {
-	return s[i].NumConfirmations < s[j].NumConfirmations
-}
-func (s byHeight) Swap(i, j int) { s[i], s[j] = s[j], s[i] }
